% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertTwelveToFourPoint.R
\name{convertTwelveToFourPoint}
\alias{convertTwelveToFourPoint}
\title{Convert your grades from the Twelve point system to the Four point system}
\usage{
convertTwelveToFourPoint(gradeVector)
}
\arguments{
\item{gradeVector}{the input is your grades as calculated on the four point scale.}
}
\value{
The return value is a vector of numbers representing your grades on the four point scale.
}
\description{
Convert your grades from the Twelve point system to the Four point system
}
\examples{
a = c(12, 12, 12, 12, 11)

convertTwelveToFourPoint(gradeVector = a)

## [1] 4.0 4.0 4.0 4.0 3.9
}
